requirements.txt
# FastAPI and server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
jinja2>=3.1.2
python-multipart>=0.0.6

# Cryptography
cryptography>=41.0.0
pynacl>=1.5.0
argon2-cffi>=23.1.0

# HTTP client
httpx>=0.25.0
aiofiles>=23.2.0

# Cerebro Care package (installed via Dockerfile with GITHUB_TOKEN)